Make the example a root tutorial and add corresponding notebook
the notebook is automatically generated with the converttonotebook
tool of the root dcoumentation set of scripts.
